Add files via upload
This project is a text prediction model that uses deep learning techniques to predict the next word given a sequence of words. The model is built using the TensorFlow and Keras libraries in Python. The dataset used to train the model is preprocessed and tokenized to convert the text data into a numerical format, which is then used to train the model. The trained model is then saved and loaded to predict the next word based on the input text.
The project includes a Python script that loads the trained model and a tokenizer and defines a function to predict the next word given an input text. The user is prompted to enter an initial text, and the model then predicts the next words based on the input text. The user can specify how many words they want to predict.
The project can be further extended by incorporating more advanced deep learning techniques such as recurrent neural networks (RNNs) and long short-term memory (LSTM) networks, which can capture more complex patterns in the text data. Additionally, the model can be trained on larger datasets to improve its accuracy and ability to predict the next word more accurately. Overall, this project demonstrates the use of deep learning techniques for text prediction, which has various applications in natural language processing (NLP) tasks.